net, l2_bridge: Move bridge_nad fixture
bridge_nad is used in vmi_interfaces_stability and
can be reused.
Moving bridge_nad to the l2_bridge package conftest
makes it available to all l2_bridge sub-packages.
