+++ tests/network/l2_bridge/vmi_interfaces_stability/conftest.py
@@ -4,8 +4,7 @@ import pytest
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.namespace import Namespace
 
-import tests.network.libs.nodenetworkconfigurationpolicy as libnncp
-from libs.net.netattachdef import CNIPluginBridgeConfig, NetConfig, NetworkAttachmentDefinition
+from libs.net.netattachdef import NetworkAttachmentDefinition
 from libs.vm.vm import BaseVirtualMachine
 from tests.network.l2_bridge.vmi_interfaces_stability.lib_helpers import (
     secondary_network_vm,
@@ -29,25 +28,6 @@ def running_linux_bridge_vm(
         vm.wait_for_agent_connected()
         wait_for_stable_ifaces(vm=vm)
         yield vm
-
-
-@pytest.fixture(scope="class")
-def bridge_nad(
-    admin_client: DynamicClient,
-    namespace: Namespace,
-    bridge_nncp: libnncp.NodeNetworkConfigurationPolicy,
-) -> Generator[NetworkAttachmentDefinition]:
-    config = NetConfig(
-        name="test-bridge-network",
-        plugins=[CNIPluginBridgeConfig(bridge=bridge_nncp.desired_state_spec.interfaces[0].name)],  # type: ignore
-    )
-    with NetworkAttachmentDefinition(
-        name="test-bridge-network",
-        namespace=namespace.name,
-        config=config,
-        client=admin_client,
-    ) as nad:
-        yield nad
 
 
 @pytest.fixture(scope="class")
